Refactor client channel tests to move all Open and event tests into a common base class
