+++ WCF/Shared.Tests/Microsoft.AI.Wcf.Tests.projitems
@@ -9,6 +9,7 @@
     <Import_RootNamespace>Microsoft.ApplicationInsights.Wcf.Tests</Import_RootNamespace>
   </PropertyGroup>
   <ItemGroup>
+    <Compile Include="$(MSBuildThisFileDirectory)ChannelTestBase.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ClientChannelManager.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ClientIpTelemetryInitializerTests.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ClientTelemetryBindingElementTests.cs" />
